+++ dags/experiments.py
@@ -16,7 +16,8 @@ import dagster
 from posthog.schema import ExperimentFunnelMetric, ExperimentMeanMetric, ExperimentQuery, ExperimentRatioMetric
 
 from posthog.hogql_queries.experiments.experiment_query_runner import ExperimentQueryRunner
-from posthog.models.experiment import Experiment, ExperimentMetricResult
+
+from products.experiments.backend.models import Experiment, ExperimentMetricResult
 
 from dags.common import JobOwners
