+++ ee/clickhouse/views/test/test_clickhouse_experiments.py
@@ -9,14 +9,14 @@ from django.core.cache import cache
 from dateutil import parser
 from rest_framework import status
 
-from posthog.models import WebExperiment
 from posthog.models.action.action import Action
 from posthog.models.activity_logging.activity_log import ActivityLog
 from posthog.models.cohort.cohort import Cohort
-from posthog.models.experiment import Experiment, ExperimentSavedMetric
 from posthog.models.feature_flag import FeatureFlag, get_feature_flags_for_team_in_cache
 from posthog.models.user import User
 from posthog.test.test_journeys import journeys_for
+
+from products.experiments.backend.models import Experiment, ExperimentSavedMetric, WebExperiment
 
 from ee.api.test.base import APILicensedTest
 from ee.clickhouse.views.experiment_saved_metrics import ExperimentToSavedMetricSerializer
