+++ ee/clickhouse/views/test/test_experiment_holdouts.py
@@ -1,7 +1,8 @@
 from rest_framework import status
 
-from posthog.models.experiment import Experiment
 from posthog.models.feature_flag import FeatureFlag
+
+from products.experiments.backend.models import Experiment
 
 from ee.api.test.base import APILicensedTest
