+++ ee/clickhouse/views/test/test_experiment_saved_metrics.py
@@ -1,6 +1,6 @@
 from rest_framework import status
 
-from posthog.models.experiment import Experiment, ExperimentToSavedMetric
+from products.experiments.backend.models import Experiment, ExperimentToSavedMetric
 
 from ee.api.test.base import APILicensedTest
