+++ posthog/admin/admins/__init__.py
@@ -8,8 +8,6 @@ from .dashboard_template_admin import DashboardTemplateAdmin
 from .data_color_theme_admin import DataColorThemeAdmin
 from .data_warehouse_table_admin import DataWarehouseTableAdmin
 from .event_ingestion_restriction_config import EventIngestionRestrictionConfigAdmin
-from .experiment_admin import ExperimentAdmin
-from .experiment_saved_metric_admin import ExperimentSavedMetricAdmin
 from .feature_flag_admin import FeatureFlagAdmin
 from .group_type_mapping_admin import GroupTypeMappingAdmin
 from .hog_function_admin import HogFunctionAdmin
@@ -36,8 +34,6 @@ __all__ = [
     "DataColorThemeAdmin",
     "DataWarehouseTableAdmin",
     "EventIngestionRestrictionConfigAdmin",
-    "ExperimentAdmin",
-    "ExperimentSavedMetricAdmin",
     "FeatureFlagAdmin",
     "GroupTypeMappingAdmin",
     "HogFunctionAdmin",
