+++ posthog/api/search.py
@@ -11,8 +11,10 @@ from rest_framework.response import Response
 
 from posthog.api.routing import TeamAndOrgViewSetMixin
 from posthog.helpers.full_text_search import build_rank, process_query
-from posthog.models import Action, Cohort, Dashboard, EventDefinition, Experiment, FeatureFlag, Insight, Survey
+from posthog.models import Action, Cohort, Dashboard, EventDefinition, FeatureFlag, Insight, Survey
 from posthog.models.notebook.notebook import Notebook
+
+from products.experiments.backend.models import Experiment
 
 LIMIT = 25
