+++ posthog/api/web_experiment.py
@@ -15,8 +15,10 @@ from posthog.api.routing import TeamAndOrgViewSetMixin
 from posthog.api.utils import get_token
 from posthog.auth import TemporaryTokenAuthentication
 from posthog.exceptions import generate_exception_response
-from posthog.models import Team, WebExperiment
+from posthog.models import Team
 from posthog.utils_cors import cors_response
+
+from products.experiments.backend.models import WebExperiment
 
 
 class WebExperimentsAPISerializer(serializers.ModelSerializer):
