+++ posthog/hogql_queries/experiments/test/test_experiment_funnels_query_runner.py
@@ -30,9 +30,10 @@ from posthog.schema import (
 
 from posthog.constants import ExperimentNoResultsErrorKeys
 from posthog.hogql_queries.experiments.experiment_funnels_query_runner import ExperimentFunnelsQueryRunner
-from posthog.models.experiment import Experiment, ExperimentHoldout
 from posthog.models.feature_flag.feature_flag import FeatureFlag
 from posthog.test.test_journeys import journeys_for
+
+from products.experiments.backend.models import Experiment, ExperimentHoldout
 
 
 class TestExperimentFunnelsQueryRunner(ClickhouseTestMixin, APIBaseTest):
