+++ posthog/models/file_system/test/test_file_system.py
@@ -1,8 +1,10 @@
 from django.test import TestCase
 
-from posthog.models import Dashboard, Experiment, FeatureFlag, Insight, Notebook, Organization, Team, User
+from posthog.models import Dashboard, FeatureFlag, Insight, Notebook, Organization, Team, User
 from posthog.models.file_system.file_system import FileSystem, escape_path, join_path, split_path
 from posthog.models.file_system.unfiled_file_saver import save_unfiled_files
+
+from products.experiments.backend.models import Experiment
 
 
 class TestFileSystemModel(TestCase):
