+++ posthog/models/file_system/unfiled_file_saver.py
@@ -7,7 +7,6 @@ from django.utils import timezone
 from posthog.models.action.action import Action
 from posthog.models.cohort import Cohort
 from posthog.models.dashboard import Dashboard
-from posthog.models.experiment import Experiment
 from posthog.models.feature_flag import FeatureFlag
 from posthog.models.file_system.file_system import FileSystem, escape_path, split_path
 from posthog.models.file_system.file_system_mixin import FileSystemSyncMixin
@@ -21,6 +20,7 @@ from posthog.models.user import User
 from posthog.session_recordings.models.session_recording_playlist import SessionRecordingPlaylist
 
 from products.early_access_features.backend.models import EarlyAccessFeature
+from products.experiments.backend.models import Experiment
 
 MIXIN_MODELS = {
     "action": Action,
