+++ posthog/models/test/test_product_intent.py
@@ -6,13 +6,14 @@ from posthog.test.base import BaseTest
 from unittest.mock import patch
 
 from posthog.models.dashboard import Dashboard
-from posthog.models.experiment import Experiment
 from posthog.models.feature_flag import FeatureFlag
 from posthog.models.insight import Insight
 from posthog.models.product_intent.product_intent import ProductIntent, calculate_product_activation
 from posthog.models.surveys.survey import Survey
 from posthog.session_recordings.models.session_recording import SessionRecording
 from posthog.utils import get_instance_realm
+
+from products.experiments.backend.models import Experiment
 
 
 class TestProductIntent(BaseTest):
