+++ posthog/settings/web.py
@@ -35,6 +35,7 @@ PRODUCTS_APPS = [
     "products.revenue_analytics",
     "products.user_interviews",
     "products.llm_analytics",
+    "products.experiments.backend.apps.ExperimentsConfig",
 ]
 
 INSTALLED_APPS = [
