+++ products/experiments/backend/models.py
@@ -1,4 +1,4 @@
-from typing import TYPE_CHECKING, Any
+from typing import Any
 
 from django.db import models
 from django.db.models import QuerySet
@@ -7,10 +7,8 @@ from django.utils import timezone
 from posthog.models.activity_logging.model_activity import ModelActivityMixin
 from posthog.models.file_system.file_system_mixin import FileSystemSyncMixin
 from posthog.models.file_system.file_system_representation import FileSystemRepresentation
-from posthog.models.utils import RootTeamMixin
-
-if TYPE_CHECKING:
-    from posthog.models.team import Team
+from posthog.models.team import Team
+from posthog.models.utils import RootTeamManager, RootTeamMixin
 
 
 class Experiment(FileSystemSyncMixin, ModelActivityMixin, RootTeamMixin, models.Model):
@@ -20,11 +18,9 @@ class Experiment(FileSystemSyncMixin, ModelActivityMixin, RootTeamMixin, models.
 
     name = models.CharField(max_length=400)
     description = models.CharField(max_length=400, null=True, blank=True)
-    team = models.ForeignKey("Team", on_delete=models.CASCADE)
-
+    team = models.ForeignKey("posthog.Team", on_delete=models.CASCADE)
     # Filters define the target metric of an Experiment
     filters = models.JSONField(default=dict, blank=True)
-
     # Parameters include configuration fields for the experiment: What the control & test variant are called,
     # and any test significance calculation parameters
     # We have 4 parameters today:
@@ -34,15 +30,12 @@ class Experiment(FileSystemSyncMixin, ModelActivityMixin, RootTeamMixin, models.
     #   feature_flag_variants: { key: string, name: string, rollout_percentage: number }[]
     #   custom_exposure_filter: Filter json
     parameters = models.JSONField(default=dict, null=True)
-
     # A list of filters for secondary metrics
     secondary_metrics = models.JSONField(default=list, null=True, blank=True)
-
-    created_by = models.ForeignKey("User", on_delete=models.SET_NULL, null=True)
-    feature_flag = models.ForeignKey("FeatureFlag", blank=False, on_delete=models.RESTRICT)
-    exposure_cohort = models.ForeignKey("Cohort", on_delete=models.SET_NULL, null=True, blank=True)
+    created_by = models.ForeignKey("posthog.User", on_delete=models.SET_NULL, null=True)
+    feature_flag = models.ForeignKey("posthog.FeatureFlag", blank=False, on_delete=models.RESTRICT)
+    exposure_cohort = models.ForeignKey("posthog.Cohort", on_delete=models.SET_NULL, null=True, blank=True)
     holdout = models.ForeignKey("ExperimentHoldout", on_delete=models.SET_NULL, null=True, blank=True)
-
     start_date = models.DateTimeField(null=True, blank=True)
     end_date = models.DateTimeField(null=True, blank=True)
     created_at = models.DateTimeField(default=timezone.now)
@@ -51,9 +44,7 @@ class Experiment(FileSystemSyncMixin, ModelActivityMixin, RootTeamMixin, models.
     deleted = models.BooleanField(default=False, null=True)
     type = models.CharField(max_length=40, choices=ExperimentType.choices, null=True, blank=True, default="product")
     variants = models.JSONField(default=dict, null=True, blank=True)
-
     exposure_criteria = models.JSONField(default=dict, null=True, blank=True)
-
     metrics = models.JSONField(default=list, null=True, blank=True)
     metrics_secondary = models.JSONField(default=list, null=True, blank=True)
     primary_metrics_ordered_uuids = models.JSONField(default=None, null=True, blank=True)
@@ -61,9 +52,7 @@ class Experiment(FileSystemSyncMixin, ModelActivityMixin, RootTeamMixin, models.
     saved_metrics: models.ManyToManyField = models.ManyToManyField(
         "ExperimentSavedMetric", blank=True, related_name="experiments", through="ExperimentToSavedMetric"
     )
-
     stats_config = models.JSONField(default=dict, null=True, blank=True)
-
     conclusion = models.CharField(
         max_length=30,
         choices=[
@@ -80,6 +69,9 @@ class Experiment(FileSystemSyncMixin, ModelActivityMixin, RootTeamMixin, models.
         null=True,
         blank=True,
     )
+
+    class Meta:
+        db_table = "posthog_experiment"
 
     def __str__(self):
         return self.name or "Untitled"
@@ -117,15 +109,16 @@ class Experiment(FileSystemSyncMixin, ModelActivityMixin, RootTeamMixin, models.
 class ExperimentHoldout(ModelActivityMixin, RootTeamMixin, models.Model):
     name = models.CharField(max_length=400)
     description = models.CharField(max_length=400, null=True, blank=True)
-    team = models.ForeignKey("Team", on_delete=models.CASCADE)
-
+    team = models.ForeignKey("posthog.Team", on_delete=models.CASCADE)
     # Filters define the definition of the holdout
     # This is then replicated across flags for experiments in the holdout
     filters = models.JSONField(default=list)
-
-    created_by = models.ForeignKey("User", on_delete=models.SET_NULL, null=True)
+    created_by = models.ForeignKey("posthog.User", on_delete=models.SET_NULL, null=True)
     created_at = models.DateTimeField(default=timezone.now)
     updated_at = models.DateTimeField(auto_now=True)
+
+    class Meta:
+        db_table = "posthog_experimentholdout"
 
     def save(self, *args: Any, skip_activity_log: bool = False, **kwargs: Any) -> None:
         if skip_activity_log:
@@ -138,29 +131,31 @@ class ExperimentHoldout(ModelActivityMixin, RootTeamMixin, models.Model):
 class ExperimentSavedMetric(ModelActivityMixin, RootTeamMixin, models.Model):
     name = models.CharField(max_length=400)
     description = models.CharField(max_length=400, null=True, blank=True)
-    team = models.ForeignKey("Team", on_delete=models.CASCADE)
-
+    team = models.ForeignKey("posthog.Team", on_delete=models.CASCADE)
     query = models.JSONField()
-
     # Metadata for the saved metric
     # has things like if this metric was migrated from a legacy metric
     metadata = models.JSONField(null=True, blank=True, default=dict)
-
-    created_by = models.ForeignKey("User", on_delete=models.SET_NULL, null=True)
+    created_by = models.ForeignKey("posthog.User", on_delete=models.SET_NULL, null=True)
     created_at = models.DateTimeField(default=timezone.now)
     updated_at = models.DateTimeField(auto_now=True)
+
+    class Meta:
+        db_table = "posthog_experimentsavedmetric"
 
 
 class ExperimentToSavedMetric(models.Model):
     experiment = models.ForeignKey("Experiment", on_delete=models.CASCADE)
     saved_metric = models.ForeignKey("ExperimentSavedMetric", on_delete=models.CASCADE)
-
     # Metadata for the saved metric at the time of the experiment creation
     # has stuff like whether this metric is primary, and any other information
     # we need for the metric, other than the query.
     metadata = models.JSONField(default=dict)
     created_at = models.DateTimeField(default=timezone.now)
     updated_at = models.DateTimeField(auto_now=True)
+
+    class Meta:
+        db_table = "posthog_experimenttosavedmetric"
 
     def __str__(self):
         return f"{self.experiment.name} - {self.saved_metric.name} - {self.metadata}"
@@ -189,6 +184,19 @@ class ExperimentMetricResult(models.Model):
         indexes = [
             models.Index(fields=["experiment", "metric_uuid", "query_to"]),
         ]
+        db_table = "posthog_experimentmetricresult"
 
     def __str__(self):
         return f"ExperimentMetricResult({self.experiment_id}, {self.metric_uuid}, {self.query_from}, {self.status})"
+
+
+class WebExperimentManager(RootTeamManager):
+    def get_queryset(self):
+        return super().get_queryset().filter(type="web")
+
+
+class WebExperiment(Experiment, RootTeamMixin):
+    objects = WebExperimentManager()  # type: ignore
+
+    class Meta:
+        proxy = True
